/reactionRoleのリファクタリング
